backend/requirements.txt
# RealStat Rental Platform — FastAPI backend dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
# pymongo version is constrained by motor; do not pin separately
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
email-validator==2.2.0
aiofiles==24.1.0
boto3==1.35.76
stripe==11.3.0
